# packages/protocol/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Spark HTTP API",
    "version": "0.1.0",
    "description": "Spark Agent 工作台的 REST+SSE API（本文件由 scripts/gen-openapi.ts 生成，勿手改；事实源 apps/server/src/routes/* 与 packages/protocol schema）。",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4318",
      "description": "本地缺省（127.0.0.1 是刻意的，ADR 红线）"
    }
  ],
...
  "paths": {
...
    "/api/usage/summary": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "成本看板（工单 13.6：总账五分量 + 明细桶 + 熔断状态）",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageSummaryDto"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "schema": {
              "type": "string",
              "description": "YYYY-MM-DD（可省 = 全量；含当日）"
            }
          }
        ]
      }
    },
    "/api/settings": {
      "get": {
        "tags": [
          "config"
        ],
        "summary": "spark.json 脱敏读（hooks 并入同一端点 = 工单 10.21）",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "config"
        ],
        "summary": "更新设置（zod 校验 → 原子写盘 → 才改内存；重启档字段带 restartRequired）",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsDto"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdate"
              }
            }
          }
        }
      }
    },
...
    "/api/sessions/{id}/commands/{name}": {
      "post": {
        "tags": [
          "config"
        ],
        "summary": "执行会话命令（client 命令 → 400 E_COMMAND_CLIENT；未知 → 404）",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "const": true
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteCommandBody"
              }
            }
          }
        }
      }
    },
    "/api/mcp": {
      "get": {
        "tags": [
          "config"
        ],
        "summary": "MCP server 连接状态快照（失败也列出 connected:false）",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/McpServerDto"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "config"
        ],
        "summary": "整文件写 mcp.json（工单 12.6：坏配置不落盘 → 400；重启后重连）",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "const": true
                    },
                    "restartRequired": {
                      "type": "boolean",
                      "const": true
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "version": {
                    "type": "integer",
                    "const": 1
                  },
                  "servers": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "object",
                      "properties": {
                        "command": {
                          "type": "string"
                        },
                        "args": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "env": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "string"
                          }
                        }
                      },
                      "required": [
                        "command"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/lsp/install": {
      "post": {
        "tags": [
          "config"
        ],
        "summary": "安装内置清单语言服务器并写入 lsp.json（阶段十九 19.5 / ADR D47；未知 id 404，npm 失败 502）",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LspInstallResultDto"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  }
                },
                "required": [
                  "id"
                ]
              }
            }
          }
        }
      }
    },
...
    "/api/memories/{id}": {
      "delete": {
        "tags": [
          "memories"
        ],
        "summary": "删除一条记忆（无此条 → 404）",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "const": true
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ]
      }
    },
...
    "/api/audit": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "审计明细流（新→旧；permission.decision / permission.rule / session.rollback）",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEntryDto"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 200
            }
          },
          {
            "name": "kind",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "permission.decision",
                "permission.rule",
                "session.rollback"
              ]
            }
          },
          {
            "name": "result",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "allow",
                "deny",
                "applied",
                "ok"
              ]
            }
          },
          {
            "name": "tool",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "since",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/api/search": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "会话全文搜索（FTS5：user/assistant 消息 + 标题；q 必填）",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchHitDto"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ]
      }
    },
    "/api/artifacts/{file}": {
      "get": {
        "tags": [
          "artifacts"
        ],
        "summary": "browser 截图供图（文件名白名单校验在引擎侧）",
        "responses": {
          "200": {
            "description": "PNG 截图字节",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "file",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditEntryDto": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "time": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "kind": {
            "type": "string",
            "enum": [
              "permission.decision",
              "permission.rule",
              "session.rollback"
            ]
          },
          "actor": {
            "type": "string",
            "enum": [
              "user",
              "system"
            ]
          },
          "result": {
            "type": "string",
            "enum": [
              "allow",
              "deny",
              "applied",
              "ok"
            ]
          },
          "sessionId": {
            "type": "string",
            "pattern": "^ses_[0-9A-Za-z]+$"
          },
          "tool": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "resource": {
            "type": "string"
          },
          "effect": {
            "type": "string",
            "enum": [
              "allow",
              "deny",
              "ask"
            ]
          },
          "op": {
            "type": "string",
            "enum": [
              "add",
              "remove"
            ]
          },
          "source": {
            "type": "string"
          },
          "checkpointId": {
            "type": "string"
          }
        },
        "required": [
          "time",
          "kind",
          "actor",
          "result"
        ],
        "additionalProperties": false
      },
...
      "ExecuteCommandBody": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "args": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "LspInstallResultDto": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "language": {
            "type": "string",
            "minLength": 1
          },
          "command": {
            "type": "string",
            "minLength": 1
          },
          "args": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "written": {
            "type": "boolean"
          }
        },
        "required": [
          "language",
          "command",
          "args",
          "written"
        ],
        "additionalProperties": false
      },
...
      "McpServerDto": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "connected": {
            "type": "boolean"
          },
          "tools": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "command": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "connected",
          "tools",
          "command"
        ],
        "additionalProperties": false
      },
...
      "SettingsDto": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "server": {
            "type": "object",
            "properties": {
              "port": {
                "type": "integer",
                "minimum": 1,
                "maximum": 65535
              },
              "host": {
                "type": "string",
                "minLength": 1
              }
            },
            "required": [
              "port",
              "host"
            ],
            "additionalProperties": false
          },
          "engine": {
            "type": "object",
            "properties": {
              "maxStepsPerTurn": {
                "type": "integer",
                "minimum": 1,
                "maximum": 9007199254740991
              },
              "maxToolParallel": {
                "type": "integer",
                "minimum": 1,
                "maximum": 9007199254740991
              },
              "toolTimeoutMs": {
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              },
              "permissionTimeoutMs": {
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              },
              "progressThrottleMs": {
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              },
              "toolOutputLimitKB": {
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              },
              "compactionThreshold": {
                "type": "number",
                "exclusiveMinimum": 0,
                "exclusiveMaximum": 1
              },
              "checkpoints": {
                "type": "boolean"
              },
              "bashSandbox": {
                "type": "string",
                "enum": [
                  "off",
                  "on"
                ]
              },
              "computerUseEnabled": {
                "type": "boolean"
              },
              "bashPersistent": {
                "type": "boolean"
              }
            },
            "required": [
              "maxStepsPerTurn",
              "maxToolParallel",
              "toolTimeoutMs",
              "permissionTimeoutMs",
              "progressThrottleMs",
              "toolOutputLimitKB",
              "compactionThreshold",
              "checkpoints",
              "bashSandbox",
              "computerUseEnabled",
              "bashPersistent"
            ],
            "additionalProperties": false
          },
          "hooks": {
            "type": "object",
            "properties": {
              "turn.before": {
                "readOnly": true,
                "type": "array",
                "items": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "command": {
                          "type": "string",
                          "minLength": 1
                        },
                        "timeoutMs": {
                          "type": "integer",
                          "exclusiveMinimum": 0,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": [
                        "command"
                      ],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "skill": {
                          "type": "string",
                          "minLength": 1
                        },
                        "emit": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "skill",
                        "emit"
                      ],
                      "additionalProperties": false
                    }
                  ]
                }
              },
              "turn.after": {
                "readOnly": true,
                "type": "array",
                "items": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "command": {
                          "type": "string",
                          "minLength": 1
                        },
                        "timeoutMs": {
                          "type": "integer",
                          "exclusiveMinimum": 0,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": [
                        "command"
                      ],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "skill": {
                          "type": "string",
                          "minLength": 1
                        },
                        "emit": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "skill",
                        "emit"
                      ],
                      "additionalProperties": false
                    }
                  ]
                }
              },
              "permission.resolved": {
                "readOnly": true,
                "type": "array",
                "items": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "command": {
                          "type": "string",
                          "minLength": 1
                        },
                        "timeoutMs": {
                          "type": "integer",
                          "exclusiveMinimum": 0,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": [
                        "command"
                      ],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "skill": {
                          "type": "string",
                          "minLength": 1
                        },
                        "emit": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "skill",
                        "emit"
                      ],
                      "additionalProperties": false
                    }
                  ]
                }
              },
              "tool.completed": {
                "readOnly": true,
                "type": "array",
                "items": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "command": {
                          "type": "string",
                          "minLength": 1
                        },
                        "timeoutMs": {
                          "type": "integer",
                          "exclusiveMinimum": 0,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": [
                        "command"
                      ],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "skill": {
                          "type": "string",
                          "minLength": 1
                        },
                        "emit": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "skill",
                        "emit"
                      ],
                      "additionalProperties": false
                    }
                  ]
                }
              }
            },
            "additionalProperties": false
          },
          "agents": {
            "type": "object",
            "properties": {
              "disabledAgents": {
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 1
                }
              }
            },
            "required": [
              "disabledAgents"
            ],
            "additionalProperties": false
          },
          "extensions": {
            "type": "object",
            "properties": {
              "disabledExtensions": {
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 1
                }
              }
            },
            "required": [
              "disabledExtensions"
            ],
            "additionalProperties": false
          },
          "restartRequired": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "models": {
            "type": "object",
            "properties": {
              "defaultModel": {
                "type": "string"
              },
              "defaultEffort": {
                "anyOf": [
                  {
                    "type": "string",
                    "enum": [
                      "low",
                      "medium",
                      "high"
                    ]
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            },
            "required": [
              "defaultModel",
              "defaultEffort"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "server",
          "engine",
          "restartRequired",
          "models"
        ],
        "additionalProperties": false
      },
...
      "UsageSummaryDto": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "total": {
            "type": "object",
            "properties": {
              "costUsd": {
                "type": "number",
                "minimum": 0
              },
              "inputTokens": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "outputTokens": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "cacheRead": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "cacheWrite": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              }
            },
            "required": [
              "costUsd",
              "inputTokens",
              "outputTokens",
              "cacheRead",
              "cacheWrite"
            ],
            "additionalProperties": false
          },
          "buckets": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "costUsd": {
                  "type": "number",
                  "minimum": 0
                },
                "inputTokens": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                },
                "outputTokens": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                },
                "cacheRead": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                },
                "cacheWrite": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                },
                "day": {
                  "type": "string",
                  "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                },
                "provider": {
                  "type": "string",
                  "minLength": 1
                },
                "model": {
                  "type": "string",
                  "minLength": 1
                }
              },
              "required": [
                "costUsd",
                "inputTokens",
                "outputTokens",
                "cacheRead",
                "cacheWrite",
                "day",
                "provider",
                "model"
              ],
              "additionalProperties": false
            }
          },
          "unbucketed": {
            "type": "object",
            "properties": {
              "costUsd": {
                "type": "number",
                "minimum": 0
              },
              "inputTokens": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "outputTokens": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "cacheRead": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "cacheWrite": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              }
            },
            "required": [
              "costUsd",
              "inputTokens",
              "outputTokens",
              "cacheRead",
              "cacheWrite"
            ],
            "additionalProperties": false
          },
          "costLimitUsd": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "exceeded": {
            "type": "boolean"
          }
        },
        "required": [
          "total",
          "buckets",
          "unbucketed",
          "costLimitUsd",
          "exceeded"
        ],
        "additionalProperties": false
      },
...
    }
  }
}